chore: use generated types from vega protos - let the cypress compile additional es6 lib
