+++ apps/trading-e2e/project.json
@@ -7,7 +7,8 @@
       "executor": "@nrwl/cypress:cypress",
       "options": {
         "cypressConfig": "apps/trading-e2e/cypress.config.js",
-        "devServerTarget": "trading:serve"
+        "devServerTarget": "trading:serve",
+        "tsConfig": "apps/trading-e2e/tsconfig.json"
       },
       "configurations": {
         "production": {
