+++ package.json
@@ -100,11 +100,14 @@
     "@babel/core": "7.12.13",
     "@babel/plugin-proposal-export-default-from": "^7.18.10",
     "@babel/preset-env": "^7.18.10",
+    "@babel/preset-react": "^7.22.5",
     "@babel/preset-typescript": "7.12.13",
     "@commitlint/cli": "^16.2.4",
     "@commitlint/config-conventional": "^16.2.4",
     "@commitlint/config-nx-scopes": "^17.4.2",
     "@cypress/grep": "^3.1.0",
+    "@cypress/webpack-batteries-included-preprocessor": "^2.4.1",
+    "@cypress/webpack-preprocessor": "^5.17.1",
     "@ethersproject/experimental": "^5.6.0",
     "@graphql-codegen/cli": "^2.11.8",
     "@graphql-codegen/near-operation-file-preset": "^2.4.1",
@@ -159,7 +162,8 @@
     "@typescript-eslint/parser": "5.35.1",
     "autoprefixer": "10.4.8",
     "babel-jest": "27.5.1",
-    "babel-loader": "8.1.0",
+    "babel-loader": "^9.1.2",
+    "css-loader": "^6.8.1",
     "cypress": "^12.9.0",
     "cypress-mochawesome-reporter": "^3.3.0",
     "cypress-real-events": "^1.8.1",
@@ -196,7 +200,9 @@
     "replace-in-file": "^6.3.2",
     "resize-observer-polyfill": "^1.5.1",
     "sass": "1.54.0",
+    "sass-loader": "^13.3.2",
     "storybook-addon-themes": "^6.1.0",
+    "style-loader": "^3.3.3",
     "tailwindcss": "^3.2.4",
     "ts-jest": "27.1.4",
     "ts-node": "10.9.1",
